fix review comments2
