+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/connectors/hive/HiveDialectQueryITCase.java
@@ -20,11 +20,16 @@ package org.apache.flink.connectors.hive;
 
 import org.apache.flink.table.HiveVersionTestUtil;
 import org.apache.flink.table.api.SqlDialect;
+import org.apache.flink.table.api.TableEnvironment;
 import org.apache.flink.table.catalog.Catalog;
 import org.apache.flink.table.catalog.GenericInMemoryCatalog;
+import org.apache.flink.table.catalog.hive.HiveCatalog;
+import org.apache.flink.table.catalog.hive.HiveTestUtils;
 import org.apache.flink.table.catalog.hive.client.HiveShim;
 import org.apache.flink.table.catalog.hive.client.HiveShimLoader;
 import org.apache.flink.table.functions.hive.conversion.HiveInspectors;
+import org.apache.flink.table.module.CoreModule;
+import org.apache.flink.table.module.hive.HiveModule;
 import org.apache.flink.table.planner.delegation.hive.HiveParserUtils;
 import org.apache.flink.types.Row;
 import org.apache.flink.util.CollectionUtil;
@@ -43,8 +48,11 @@ import org.apache.hadoop.hive.serde2.objectinspector.primitive.PrimitiveObjectIn
 import org.apache.hadoop.hive.serde2.objectinspector.primitive.PrimitiveObjectInspectorUtils;
 import org.apache.hadoop.hive.serde2.objectinspector.primitive.TimestampObjectInspector;
 import org.apache.hadoop.hive.serde2.typeinfo.TypeInfoFactory;
+import org.junit.BeforeClass;
+import org.junit.ClassRule;
 import org.junit.ComparisonFailure;
 import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -65,7 +73,101 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /** Test hive query compatibility. */
-public class HiveDialectQueryITCase extends HiveDialectITCaseBase {
+public class HiveDialectQueryITCase {
+
+    @ClassRule public static TemporaryFolder tempFolder = new TemporaryFolder();
+
+    private static final String QTEST_DIR =
+            Thread.currentThread().getContextClassLoader().getResource("query-test").getPath();
+    private static final String SORT_QUERY_RESULTS = "SORT_QUERY_RESULTS";
+
+    private static HiveCatalog hiveCatalog;
+    private static TableEnvironment tableEnv;
+    private static String warehouse;
+
+    @BeforeClass
+    public static void setup() throws Exception {
+        hiveCatalog = HiveTestUtils.createHiveCatalog();
+        // required by query like "src.`[k].*` from src"
+        hiveCatalog.getHiveConf().setVar(HiveConf.ConfVars.HIVE_QUOTEDID_SUPPORT, "none");
+        hiveCatalog.open();
+        tableEnv = getTableEnvWithHiveCatalog();
+        warehouse = hiveCatalog.getHiveConf().getVar(HiveConf.ConfVars.METASTOREWAREHOUSE);
+
+        // create tables
+        tableEnv.executeSql("create table foo (x int, y int)");
+        tableEnv.executeSql("create table bar(I int, s string)");
+        tableEnv.executeSql("create table baz(ai array<int>, d double)");
+        tableEnv.executeSql(
+                "create table employee(id int,name string,dep string,salary int,age int)");
+        tableEnv.executeSql("create table dest (x int, y int)");
+        tableEnv.executeSql("create table destp (x int) partitioned by (p string, q string)");
+        tableEnv.executeSql("alter table destp add partition (p='-1',q='-1')");
+        tableEnv.executeSql("CREATE TABLE src (key STRING, value STRING)");
+        tableEnv.executeSql("CREATE TABLE t_sub_query (x int)");
+        tableEnv.executeSql(
+                "CREATE TABLE srcpart (key STRING, `value` STRING) PARTITIONED BY (ds STRING, hr STRING)");
+        tableEnv.executeSql("create table binary_t (a int, ab array<binary>)");
+
+        tableEnv.executeSql(
+                "CREATE TABLE nested (\n"
+                        + "  a int,\n"
+                        + "  s1 struct<f1: boolean, f2: string, f3: struct<f4: int, f5: double>, f6: int>,\n"
+                        + "  s2 struct<f7: string, f8: struct<f9 : boolean, f10: array<int>, f11: map<string, boolean>>>,\n"
+                        + "  s3 struct<f12: array<struct<f13:string, f14:int>>>,\n"
+                        + "  s4 map<string, struct<f15:int>>,\n"
+                        + "  s5 struct<f16: array<struct<f17:string, f18:struct<f19:int>>>>,\n"
+                        + "  s6 map<string, struct<f20:array<struct<f21:struct<f22:int>>>>>\n"
+                        + ")");
+        HiveTestUtils.createTextTableInserter(hiveCatalog, "default", "foo")
+                .addRow(new Object[] {1, 1})
+                .addRow(new Object[] {2, 2})
+                .addRow(new Object[] {3, 3})
+                .addRow(new Object[] {4, 4})
+                .addRow(new Object[] {5, 5})
+                .commit();
+        HiveTestUtils.createTextTableInserter(hiveCatalog, "default", "bar")
+                .addRow(new Object[] {1, "a"})
+                .addRow(new Object[] {1, "aa"})
+                .addRow(new Object[] {2, "b"})
+                .commit();
+        HiveTestUtils.createTextTableInserter(hiveCatalog, "default", "baz")
+                .addRow(new Object[] {Arrays.asList(1, 2, 3), 3.0})
+                .commit();
+        HiveTestUtils.createTextTableInserter(hiveCatalog, "default", "src")
+                .addRow(new Object[] {"1", "val1"})
+                .addRow(new Object[] {"2", "val2"})
+                .addRow(new Object[] {"3", "val3"})
+                .commit();
+        HiveTestUtils.createTextTableInserter(hiveCatalog, "default", "t_sub_query")
+                .addRow(new Object[] {2})
+                .addRow(new Object[] {3})
+                .commit();
+        HiveTestUtils.createTextTableInserter(hiveCatalog, "default", "employee")
+                .addRow(new Object[] {1, "A", "Management", 4500, 55})
+                .addRow(new Object[] {2, "B", "Management", 4400, 61})
+                .addRow(new Object[] {3, "C", "Management", 4000, 42})
+                .addRow(new Object[] {4, "D", "Production", 3700, 35})
+                .addRow(new Object[] {5, "E", "Production", 3500, 24})
+                .addRow(new Object[] {6, "F", "Production", 3600, 28})
+                .addRow(new Object[] {7, "G", "Production", 3800, 35})
+                .addRow(new Object[] {8, "H", "Production", 4000, 52})
+                .addRow(new Object[] {9, "I", "Service", 4100, 40})
+                .addRow(new Object[] {10, "J", "Sales", 4300, 36})
+                .addRow(new Object[] {11, "K", "Sales", 4100, 38})
+                .commit();
+
+        // create functions
+        tableEnv.executeSql(
+                "create function hiveudf as 'org.apache.hadoop.hive.ql.udf.generic.GenericUDFAbs'");
+        tableEnv.executeSql(
+                "create function hiveudtf as 'org.apache.hadoop.hive.ql.udf.generic.GenericUDTFExplode'");
+        tableEnv.executeSql("create function myudtf as '" + MyUDTF.class.getName() + "'");
+
+        // create temp functions
+        tableEnv.executeSql(
+                "create temporary function temp_abs as 'org.apache.hadoop.hive.ql.udf.generic.GenericUDFAbs'");
+    }
 
     @Test
     public void testQueries() throws Exception {
@@ -875,6 +977,96 @@ public class HiveDialectQueryITCase extends HiveDialectITCaseBase {
         }
     }
 
+    @Test
+    public void testSumAggFunctionPlan() {
+        // test explain
+        String actualPlan = explainSql("select x, sum(y) from foo group by x");
+        assertThat(actualPlan).isEqualTo(readFromResource("/explain/testSumAggFunctionPlan.out"));
+    }
+
+    @Test
+    public void testSimpleSumAggFunction() throws Exception {
+        tableEnv.executeSql(
+                "create table test(x string, y string, z int, d decimal(10,5), e float, f double, ts timestamp)");
+        tableEnv.executeSql(
+                        "insert into test values (NULL, '2', 1, 1.11, 1.2, 1.3, '2021-08-04 16:26:33.4'), "
+                                + "(NULL, 'b', 2, 2.22, 2.3, 2.4, '2021-08-05 16:26:33.4'), "
+                                + "(NULL, '4', 3, 3.33, 3.5, 3.6, '2021-08-07 16:26:33.4'), "
+                                + "(NULL, NULL, 4, 4.45, 4.7, 4.8, '2021-08-09 16:26:33.4')")
+                .await();
+
+        // test sum with all elements are null
+        List<Row> result =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql("select sum(x) from test").collect());
+        assertThat(result.toString()).isEqualTo("[+I[null]]");
+
+        // test sum string type with partial element can't convert to double, result type is double
+        List<Row> result2 =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql("select sum(y) from test").collect());
+        assertThat(result2.toString()).isEqualTo("[+I[6.0]]");
+
+        // test decimal type
+        List<Row> result3 =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql("select sum(d) from test").collect());
+        assertThat(result3.toString()).isEqualTo("[+I[11.11000]]");
+
+        // test sum int, result type is bigint
+        List<Row> result4 =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql("select sum(z) from test").collect());
+        assertThat(result4.toString()).isEqualTo("[+I[10]]");
+
+        // test float type
+        List<Row> result5 =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql("select sum(e) from test").collect());
+        float actualFloatValue = ((Double) result5.get(0).getField(0)).floatValue();
+        assertThat(actualFloatValue).isEqualTo(11.7f);
+
+        // test double type
+        List<Row> result6 =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql("select sum(f) from test").collect());
+        actualFloatValue = ((Double) result6.get(0).getField(0)).floatValue();
+        assertThat(actualFloatValue).isEqualTo(12.1f);
+
+        // test timestamp type
+        List<Row> result7 =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql("select sum(ts) from test").collect());
+        assertThat(result7.toString()).isEqualTo("[+I[6.513039972E9]]");
+
+        // test sum string&int type simultaneously
+        List<Row> result8 =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql("select sum(y), sum(z) from test").collect());
+        assertThat(result8.toString()).isEqualTo("[+I[6.0, 10]]");
+
+        tableEnv.executeSql("drop table test");
+    }
+
+    @Test
+    public void testSumAggWithGroupKey() throws Exception {
+        tableEnv.executeSql("create table test(name string, num bigint, price decimal(10,5))");
+        tableEnv.executeSql(
+                        "insert into test values ('tom', 2, 7.2), ('tony', 2, 23.7), ('tom', 10, 3.33), ('tony', 4, 4.45), ('nadal', 4, 10.455)")
+                .await();
+
+        List<Row> result =
+                CollectionUtil.iteratorToList(
+                        tableEnv.executeSql(
+                                        "select name, sum(num), sum(price),  sum(num * price) from test group by name")
+                                .collect());
+        assertThat(result.toString())
+                .isEqualTo(
+                        "[+I[tom, 12, 10.53000, 47.70000], +I[tony, 6, 28.15000, 65.20000], +I[nadal, 4, 10.45500, 41.82000]]");
+
+        tableEnv.executeSql("drop table test");
+    }
+
     private void runQFile(File qfile) throws Exception {
         QTest qTest = extractQTest(qfile);
         for (int i = 0; i < qTest.statements.size(); i++) {
@@ -962,6 +1154,37 @@ public class HiveDialectQueryITCase extends HiveDialectITCaseBase {
             }
         }
         return new QTest(sqlStatements, results, sortResults);
+    }
+
+    private void runQuery(String query) throws Exception {
+        try {
+            CollectionUtil.iteratorToList(tableEnv.executeSql(query).collect());
+        } catch (Exception e) {
+            System.out.println("Failed to run " + query);
+            throw e;
+        }
+    }
+
+    private String explainSql(String sql) {
+        return (String)
+                CollectionUtil.iteratorToList(tableEnv.executeSql("explain " + sql).collect())
+                        .get(0)
+                        .getField(0);
+    }
+
+    private static TableEnvironment getTableEnvWithHiveCatalog() {
+        TableEnvironment tableEnv = HiveTestUtils.createTableEnvInBatchMode(SqlDialect.HIVE);
+        tableEnv.registerCatalog(hiveCatalog.getName(), hiveCatalog);
+        tableEnv.useCatalog(hiveCatalog.getName());
+        // automatically load hive module in hive-compatible mode
+        HiveModule hiveModule = new HiveModule(hiveCatalog.getHiveVersion());
+        CoreModule coreModule = CoreModule.INSTANCE;
+        for (String loaded : tableEnv.listModules()) {
+            tableEnv.unloadModule(loaded);
+        }
+        tableEnv.loadModule("hive", hiveModule);
+        tableEnv.loadModule("core", coreModule);
+        return tableEnv;
     }
 
     /** A test UDTF that takes multiple parameters. */
